requirement.txt
Flask-Mail==0.9.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.0.2
requests==2.28.1
SQLAlchemy==1.4.45